Move common and utils to src/dive/{common,utils}
