+++ layer/CMakeLists.txt
@@ -51,7 +51,7 @@ add_library(${target_name} SHARED ${HDR_FILES} ${SRC_FILES})
 
 add_definitions(-DLAYERNAME=Dive)
 
-set(TARGET_LINK_LIBS PRIVATE service)
+set(TARGET_LINK_LIBS PRIVATE dive_legacy_includes service)
 
 add_definitions(-DVK_USE_PLATFORM_ANDROID_KHR)
 list(
