+++ scripts/format_cmake.bat
@@ -30,7 +30,6 @@ plugins ^
 runtime_layer ^
 trace_stats ^
 ui ^
-utils ^
 src
 
 gersemi -i --definitions %SRC_DIRS% --indent 4 CMakeLists.txt %SRC_DIRS%
