+++ src/dive/common/log.h
