+++ src/dive/common/macros.h
@@ -18,6 +18,7 @@ limitations under the License.
 
 #include <string>
 #include <thread>
+
 #include "absl/status/status.h"
 #include "absl/status/statusor.h"
