+++ src/dive/utils/version_info.h
@@ -17,6 +17,7 @@ limitations under the License.
 #pragma once
 
 #include <string>
+
 namespace Dive
 {
